requirements.txt
# Core dependencies
pymongo==4.6.1
python-dotenv==1.0.0
pyyaml==6.0.1

# API dependencies
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
mongomock==4.1.2 